requirements.txt
# Django & DRF
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0

# Production WSGI server
gunicorn>=21.2,<23.0

# Database
psycopg2-binary>=2.9,<3.0

# Celery & Broker
celery>=5.3,<6.0
redis>=5.0,<6.0

# Data Processing
pandas>=2.0,<3.0
openpyxl>=3.1,<4.0

# Date Utilities
python-dateutil>=2.8,<3.0

# Environment Variables
python-dotenv>=1.0,<2.0

# Testing
coverage>=7.0,<8.0
factory-boy>=3.3,<4.0
